+++ source/Internal/Module/Configuration/DataObject/ShopConfiguration.php
@@ -7,7 +7,7 @@
 
 namespace OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject;
 
-use DomainException;
+use OxidEsales\EshopCommunity\Internal\Adapter\Exception\ModuleConfigurationNotFoundException;
 
 /**
  * @internal
@@ -34,16 +34,15 @@ class ShopConfiguration
     /**
      * @param string $moduleId
      *
-     * @throws DomainException
-     *
      * @return ModuleConfiguration
+     * @throws ModuleConfigurationNotFoundException
      */
     public function getModuleConfiguration(string $moduleId): ModuleConfiguration
     {
         if (array_key_exists($moduleId, $this->moduleConfigurations)) {
             return $this->moduleConfigurations[$moduleId];
         }
-        throw new DomainException('There is no module configuration with id ' . $moduleId);
+        throw new ModuleConfigurationNotFoundException('There is no module configuration with id ' . $moduleId);
     }
 
     /**
@@ -68,14 +67,14 @@ class ShopConfiguration
     /**
      * @param string $moduleId
      *
-     * @throws DomainException
+     * @throws ModuleConfigurationNotFoundException
      */
     public function deleteModuleConfiguration(string $moduleId)
     {
         if (array_key_exists($moduleId, $this->moduleConfigurations)) {
             unset($this->moduleConfigurations[$moduleId]);
         } else {
-            throw new DomainException('There is no module configuration with id ' . $moduleId);
+            throw new ModuleConfigurationNotFoundException('There is no module configuration with id ' . $moduleId);
         }
     }