+++ tests/Acceptance/Admin/ModuleTest.php
@@ -9,6 +9,11 @@ namespace OxidEsales\EshopCommunity\Tests\Acceptance\Admin;
 use OxidEsales\Eshop\Application\Controller\ContentController;
 use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\EshopCommunity\Internal\Application\ContainerFactory;
+use OxidEsales\EshopCommunity\Internal\Module\Configuration\Bridge\ModuleConfigurationDaoBridgeInterface;
+use OxidEsales\EshopCommunity\Internal\Module\Configuration\Bridge\ShopConfigurationDaoBridgeInterface;
+use OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject\ModuleConfiguration;
+use OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject\ModuleSetting;
+use Webmozart\PathUtil\Path;
 
 /**
  * Admin interface functionality.
@@ -288,86 +293,6 @@ class ModuleTest extends ModuleBaseTest
         }
 
         $this->loginAdmin("Extensions", "Modules");
-        $this->openListItem("Test module #6 (in vendor dir)");
-        $this->clickAndWait("//form[@id='myedit']//input[@value='Activate']");
-        $this->assertElementPresent("//form[@id='myedit']//input[@value='Deactivate']");
-        $this->assertTextPresent("1.0");
-        $this->assertTextPresent("OXID eSales");
-        $this->assertTextPresent("info@oxid-esales.com");
-        $this->assertTextPresent("http://www.oxid-esales.com");
-
-        $this->openListItem("Test module #6 (in vendor dir)");
-        $this->openTab("Test tab EN");
-
-        $this->clearCache();
-        $this->openShop();
-        $this->open(shopURL."en/About-Us/");
-        $this->assertTextPresent("About Us + info6");
-
-        // vendor file name is change from myinfo6 to myinfo6test
-        $this->updateInformationAboutShopExtension('oxid/test6/view/myinfo6', 'oxid/test6/view/myinfo6test');
-
-        $this->openShop();
-        $this->open(shopURL."en/About-Us/");
-        $this->assertTextNotPresent(strtoupper("About Us + info6"));
-        $this->clearCache();
-        $this->openShop();
-        $this->assertTextNotPresent("Module #6 title EN");
-        $this->switchLanguage("Deutsch");
-        $this->assertTextNotPresent("Module #6 title DE");
-
-        // Check if error logged
-        $this->assertLoggedException(\OxidEsales\Eshop\Core\Exception\SystemComponentException::class, "Module class myinfo6test not found");
-
-        //checking if module is deactive after  vendor file rename
-        $this->loginAdmin("Extensions", "Modules");
-        $this->frame("edit");
-        $this->assertTextPresent("Invalid modules were detected");
-        $this->assertTextPresent("Do you want to delete all registered module information and saved configurations");
-        $this->assertTextPresent("OxidEsales\\Eshop\\Application\\Controller\\ContentController => oxid/test6/view/myinfo6test");
-        $this->clickAndWaitFrame("yesButton");
-        $this->openListItem("link=Test module #6 (in vendor dir)");
-        $this->assertElementNotPresent("//form[@id='myedit']//input[@value='Deactivate']");
-
-        //checking if module (oxblock, menu.xml) is disabled in shop after vendor file rename
-        $this->clearCache();
-        $this->openShop();
-        $this->assertTextNotPresent("Module #6 title EN");
-        $this->switchLanguage("Deutsch");
-        $this->assertTextNotPresent("Module #6 title DE");
-        $this->clearCache();
-        $this->openShop();
-        $this->open(shopURL."en/About-Us/");
-        $this->assertTextNotPresent("About Us + info6");
-
-        //file name is reset to the original
-        $aModules = array('content' => 'oxid/test6/view/myinfo6');
-        Registry::getConfig()->saveShopConfVar("aarr", "aModules", $aModules);
-
-        $this->loginAdmin("Extensions", "Modules");
-        $this->clickAndWait("link=Test module #6 (in vendor dir)");
-        $this->frame("edit");
-        $this->clickAndWait("//form[@id='myedit']//input[@value='Activate']");
-        $this->assertElementPresent("//form[@id='myedit']//input[@value='Deactivate']");
-        $this->frame('list');
-        $this->waitForElement("link=Test tab EN");
-        $this->openTab("Test tab EN");
-        $this->assertTextPresent("oxid/test6/view/myinfo6");
-    }
-
-    /**
-     * Testing modules in vendor directory. Checking when any file with source code class of module is deleted.
-     *
-     * @group adminFunctionality
-     */
-    public function testModulesHandlingDeleteVendorDir()
-    {
-        $testConfig = $this->getTestConfig();
-        if ($testConfig->isSubShop()) {
-            $this->markTestSkipped("Test is not for SubShop");
-        }
-
-        $this->loginAdmin("Extensions", "Modules");
         $this->clickAndWait("link=Test module #6 (in vendor dir)");
         $this->frame("edit");
         $this->clickAndWait("//form[@id='myedit']//input[@value='Activate']");
@@ -380,8 +305,12 @@ class ModuleTest extends ModuleBaseTest
         $this->open(shopURL."en/About-Us/");
         $this->assertTextPresent("About Us + info6");
 
-        //vendor dir name is changed from test6 to test6test
-        $this->updateInformationAboutShopExtension('oxid/test6/view/myinfo6', 'oxid/test6/view/deleted');
+        $moduleDirectoryPath= $this->getTestConfig()->getShopPath() . DIRECTORY_SEPARATOR . 'modules' . DIRECTORY_SEPARATOR ;
+
+        rename(
+            $moduleDirectoryPath . 'oxid/test6/view/myinfo6.php',
+            $moduleDirectoryPath . 'oxid/test6/view/deleted'
+        );
 
         $this->openShop();
         $this->open(shopURL."en/About-Us/");
@@ -417,8 +346,11 @@ class ModuleTest extends ModuleBaseTest
         $this->open(shopURL."en/About-Us/");
         $this->assertTextNotPresent("About Us + info6");
 
-        //checking if is restore the vendor dir name to original
-        Registry::getConfig()->saveShopConfVar("aarr", "aModules", array());
+        rename(
+            $moduleDirectoryPath . 'oxid/test6/view/deleted',
+            $moduleDirectoryPath . 'oxid/test6/view/myinfo6.php'
+        );
+
         $this->loginAdmin("Extensions", "Modules");
         $this->clickAndWait("link=Test module #6 (in vendor dir)");
         $this->clickAndWait("link=Test module #6 (in vendor dir)");
@@ -429,25 +361,5 @@ class ModuleTest extends ModuleBaseTest
         $this->waitForElement("link=Test tab EN");
         $this->openTab("Test tab EN");
         $this->assertTextPresent("oxid/test6/view/myinfo6");
-    }
-
-    /**
-     * Change module id in chain to imitate that module was updated in file system.
-     *
-     * @param string $existingExtensionInformation
-     * @param string $newExtensionInformation
-     */
-    private function updateInformationAboutShopExtension($existingExtensionInformation, $newExtensionInformation)
-    {
-        Registry::set(\OxidEsales\Eshop\Core\Config::class, null);
-        $activeModules = Registry::getConfig()->getConfigParam("aModules");
-        foreach ($activeModules as $shopClassName => $moduleClassName) {
-            if (strpos($moduleClassName, $existingExtensionInformation) !== false) {
-                $updatedModuleClassName = str_replace($existingExtensionInformation, $newExtensionInformation, $moduleClassName);
-                $activeModules[$shopClassName] = $updatedModuleClassName;
-            }
-        }
-        Registry::getConfig()->saveShopConfVar("aarr", "aModules", $activeModules);
-        $this->clearCache();
     }
 }