+++ tests/Unit/Internal/Module/Configuration/DataObject/ShopConfigurationTest.php
@@ -8,7 +8,7 @@ declare(strict_types = 1);
 
 namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Module\Configuration\DataObject;
 
-use DomainException;
+use OxidEsales\EshopCommunity\Internal\Adapter\Exception\ModuleConfigurationNotFoundException;
 use OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject\ClassExtensionsChain;
 use OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject\ModuleConfiguration;
 use OxidEsales\EshopCommunity\Internal\Module\Configuration\DataObject\ShopConfiguration;
@@ -73,13 +73,13 @@ class ShopConfigurationTest extends TestCase
 
     public function testGetModuleConfigurationThrowsExceptionIfModuleIdNotPresent()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(ModuleConfigurationNotFoundException::class);
         $this->shopConfiguration->getModuleConfiguration('moduleIdNotPresent');
     }
 
     public function testDeleteModuleConfigurationThrowsExceptionIfModuleIdNotPresent()
     {
-        $this->expectException(DomainException::class);
+        $this->expectException(ModuleConfigurationNotFoundException::class);
         $this->shopConfiguration->deleteModuleConfiguration('moduleIdNotPresent');
     }